Increase the heap
The article is wrong:
https://www.thewindowsclub.com/out-of-memory-error-while-copying-files

It's a) not a DWORD b) it doesn't exist by default c) you not fix anything here you increase the heap d) you not need to touch this! 

Correct article written by me several years ago (XP times)
https://ss64.com/nt/syntax-desktopheap.html

DO NOT USE THIS! "Tweak" it's not needed!