Added a Wikipedia link which explains all WIndows 10 versions (Pro/Home/ENt/LTSC) and what they contain
Because I often get asked "what is the best version". Decide yourself. 

PS: 
LTSC people are to 99,9% pirates and MS knows that cause there is only an official 90 days trial avbl. for download and the "full" version requires a Ent. Membership. No one is going to buy a 500+ Dollar account each year just for a LTSC Edition which is basically a slimmed down Ent. version.