Workaround for 0x800705b4 (KB4495666)
By default, no value is set unless you manually opt-in into the Sandbox or Windows Defender Application Guard feature (part of the Sandbox upper layer). it has no effect if the values are stored and set to 0 since the user has full control over the toggles, that's why there is no "undo".