Security baseline (DRAFT) for Windows 10 v1903 and Windows Server v1903
Why is this important or gets under research?
Well, basically spoken the "security migration" mentioned or listed in the draft or in the final version might get merged (depends if they get my approval) - not everything MS "recommend" is good e.g. LLMNR "hardening" took years to make it in the list which I mentioned it since Windows XP times and now it finally made it into the list. 

However, there are drawbacks:
- GPO and even sometimes regedit is not reliable because several factors, GPO can and still is bugged, AV programs tend to override toggles or other "tweak utilities" can interfere into this.
- Some new "tweaks" are not needed in case you use local accounts only (it makes no sense to harden something which never gets used because that might only ends-up with problems). 
- Some mentioned stuff (still) can be bypassed (I'm not pointing out every exploit which still works in 1903).