Re-enable function and quality updates settings for upgrade users
This will bring back the advance settings under Windows Update to control function-/quality updates.

MS is aware that this is currently "broken" and might fix it with the next update. However, the "solution" might or might not work for everyone because the Windows Update GPO settings automatically overriding it (you see grayed-out settings) also keep in mind that this will not affect users which fresh installed Windows 1903 (May Update).