nVidia: Enable middle mouse push-to-talk button on ShadowPlay
Does not work with 3.x SP because Windows handles the shortcuts: 
In case  there is no option in SP 3.x to assign PTT to your mouse then this basically means the mouse driver doesn't support it. SP 3.x checks Windows and then the Mouse itself for shortcut handlers, that's by design and one of a "security" features in Windows 10.