Show/Hide Chrome Bookmark Bar (also works in New Page Tab (NTP))
This will automatically trigger "Your Browser is Managed by Your Organization" in case you want to workaround this, go into chrome://flags/ and search for: "Show managed UI for managed users". Disable it and restart Chrome.

You can also remove the "Your Browser is Managed by Your Organization" function via registry but Chrome will overwrite it once the chome.exe process is restarted, so my advice is to workaround this with a flag.