Improve Surface's touch responsiveness
This tweak is (officially) placebo, however some people claim it helps to reduce the inking "latency". I did not tested it, so you are on your own. 

In 19H2 this is not needed anymore (see link within the reg file). However, the values are still there, so I decided to put it under the "All Versions" section.